feat: make react-native-share and react-native-haptic-feedback optional (#1744)
