+++ package/native-package/src/handlers/Video.tsx
@@ -1,0 +1,22 @@
+import AudioVideoPlayer from '../optionalDependencies/Video';
+export const Video = AudioVideoPlayer
+  ? ({ onBuffer, onEnd, onLoad, onProgress, paused, repeat, style, uri, videoRef }) => (
+      <AudioVideoPlayer
+        ignoreSilentSwitch={'ignore'}
+        onBuffer={onBuffer}
+        onEnd={onEnd}
+        onError={(error) => {
+          console.error(error);
+        }}
+        onLoad={onLoad}
+        onProgress={onProgress}
+        paused={paused}
+        ref={videoRef}
+        repeat={repeat}
+        source={{
+          uri,
+        }}
+        style={style}
+      />
+    )
+  : null;
